Removed Output Objects.
I hate Java so much. I cannot reiterate this enough. Java is useless when it comes to creating libraries that can come alive. To create a map, you cannot use brackets and set the value, you have to use a put function. You also cannot override symbols like +, so it is stupid for libraries where special objects are in existance.

Signed-off-by: William-McGonagle <mcgonaglew@gfacademy.org>